requirements.txt
tqdm
colorama
pytesseract
pillow
PyMuPDF
faiss-cpu
ollama